moved scripts to \scripts
